登录页面及删除位置的修改

+++ src/main/resources/static/js/member.js
@@ -19,6 +19,7 @@ layui.use(['table','form','layer'], function() {
                     return d.mpwd=='*'?'**********':'**********';}}
                 ,{field: 'type', title: '权限',width: 250, templet:function(d){
                     return d.type=='0'?'普通管理员':'超级管理员';}}
+                , {title: '操作', align: 'center', width: 80, toolbar: '#member'}
             ]
         ]
         , id: 'tableOne'
@@ -122,34 +123,52 @@ function addMember() {
     }
 }
 
+// 删除用户
 function delMember() {
-    $.ajax({
-        url: '/manager',
-        type: 'delete',
-        headers: {
-            'wrnm': localStorage.wrnm
-        },
-        data: {
-            "name": $("#mname").val(),
-        },
-        success: function (res) {
-            if (res.code == 200) {
-                alert("删除成功");
-            } else {
-                if (res.message !== "") {
-                    alert(res.message);
+    layui.use('layer', function () {
+        var $ = layui.jquery;
+        // 删除操作
+        layui.use(['table'], function () {
+            var table = layui.table;
+            table.on('tool(member)', function (obj) {
+                var tr = obj.data;
+                var msg = "您真的确定要删除吗？";
+                if (confirm(msg) === true) {
+                    $.ajax({
+                        url: '/manager',
+                        type: 'delete',
+                        headers: {
+                            'wrnm': localStorage.wrnm
+                        },
+                        data: {
+                            "name": tr.mname
+                        },
+                        success: function (res) {
+                            if (res.code == 200) {
+                                alert("删除成功");
+                                location.reload();
+                            } else {
+                                if (res.message !== "") {
+                                    alert(res.message);
+                                    location.reload();
+                                }
+                                else {
+                                    alert("出现异常，请重试");
+                                }
+                            }
+                        },
+                        error: function (res) {
+                            if (res.responseJSON === "") {
+                                alert("出现异常，请重试");
+                            } else {
+                                alert(res.responseJSON.message);
+                            }
+                        }
+                    });
+                } else {
+                    return false;
                 }
-                else {
-                    alert("出现异常，请重试");
-                }
-            }
-        },
-        error: function (res) {
-            if (res.responseJSON === "") {
-                alert("出现异常，请重试");
-            } else {
-                alert(res.responseJSON.message);
-            }
-        }
-    });
+            })
+        });
+    })
 }
